+++ src/objects/data/game/card.data.ts
